FIX distribute.sh, only delete pthreads.dll when building under mingw
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@7534 8aca7d54-2c30-db11-9de9-000461428c89

+++ distribute.sh
@@ -126,5 +126,8 @@ distribute
 
 # .. finally delete executable and language files
 rm simutrans/simutrans$simexe
-rm simutrans/pthread*.dll
-#rm simutrans/text/*.tab
+
+# cleanup dll's
+if [ "$OST" = "mingw" ]; then
+	rm simutrans/pthread*.dll
+fi
